requirements.txt
streamlit~=1.34.0
pillow~=10.3.0